requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
mysqlclient==2.2.6
numpy==2.0.2
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
